[DOC] Add documentation and hooks.

+++ arrow/download.sh
@@ -47,8 +47,9 @@ export -f download
 
 awk '{print $2}' arrow/thirdparty.list | xargs -P$(nproc) -I{} bash -c "download $CACHE_DIR {}"
 
-wget -nv -O $CACHE_DIR/src.tar $ARROW_CACHE_URL_PREFIX/src.tar 2>/dev/null || true
-if [[ -f $CACHE_DIR/src.tar ]] && [[ -s $CACHE_DIR/src.tar ]]; then
-  tar -xf $CACHE_DIR/src.tar -C arrow/src
+if [[ ! -f arrow/src/cpp/CMakeLists.txt ]]; then
+  wget -nv -O $CACHE_DIR/src.tar $ARROW_CACHE_URL_PREFIX/src.tar 2>/dev/null || true
+  if [[ -f $CACHE_DIR/src.tar ]] && [[ -s $CACHE_DIR/src.tar ]]; then
+    tar -xf $CACHE_DIR/src.tar -C arrow/src
+  fi
 fi
-
